Separate edit and train controls
